+++ build-logic/src/main/java/me/champeau/astro4j/DocCategory.java
@@ -11,5 +11,6 @@ public enum DocCategory {
     ANALYSIS,
     DRAW,
     ANIMATION,
-    STACKING
+    STACKING,
+    CONTROL_STRUCTURE
 }